refact: 이미지 압축 함수 분리
ECOPERCENT-245
