+++ src/Components/Home/Profile/ProfileEditor/ProfileImg.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import imageCompression from "browser-image-compression";
+import { imgCompress } from "../../../../Utils/convert";
 import * as S from "./style";
 
 export default function ProfileImg({
@@ -8,24 +8,7 @@ export default function ProfileImg({
   preview,
   setPreview,
 }) {
-  async function imgCompress(img) {
-    const options = {
-      maxSizeMB: 0.2,
-      maxWidthOrHeight: 1920,
-      useWebWorker: true,
-    };
-
-    try {
-      const compressdImg = await imageCompression(img, options);
-      return compressdImg;
-    } catch (error) {
-      // TODO: 이미지 압축 실패 시 핸들링
-      console.log("이미지 압축 실패", error);
-      return img;
-    }
-  }
-
-  async function onUpload(e) {
+  const onUpload = async (e) => {
     const uploadedImg = e.target.files[0];
     const reader = new FileReader();
     const compressedImg = await imgCompress(uploadedImg);
@@ -35,7 +18,7 @@ export default function ProfileImg({
     reader.onloadend = () => {
       setPreview(reader.result);
     };
-  }
+  };
 
   return (
     <form>
@@ -55,9 +38,7 @@ export default function ProfileImg({
         id="profile-imgFile-input"
         type="file"
         accept="image/*"
-        onChange={(e) => {
-          return onUpload(e);
-        }}
+        onChange={onUpload}
       />
     </form>
   );
