+++ src/Utils/convert.jsx
@@ -1,4 +1,6 @@
-export default function b64toBlob(b64Data, contentType) {
+import imageCompression from "browser-image-compression";
+
+export function b64toBlob(b64Data, contentType) {
   const sliceSize = 512;
 
   const byteCharacters = atob(b64Data);
@@ -16,4 +18,21 @@ export default function b64toBlob(b64Data, contentType) {
     byteArrays.push(byteArray);
   }
   return new File(byteArrays, "pot", { type: contentType });
+}
+
+export async function imgCompress(img) {
+  const options = {
+    maxSizeMB: 0.2,
+    maxWidthOrHeight: 1920,
+    useWebWorker: true,
+  };
+
+  try {
+    const compressdImg = await imageCompression(img, options);
+    return compressdImg;
+  } catch (error) {
+    // TODO: 이미지 압축 실패 시 핸들링
+    console.log("이미지 압축 실패", error);
+    return img;
+  }
 }
